Upgraded TDLib to tdlib/td@e44418f + Migrate to 64-bit file sizes + View TDLib source code in `Settings` + Improved `Copy Report Details` contents + Minor improvements + Version bump to `1533`

+++ app/src/main/java/org/thunderdog/challegram/data/MediaWrapper.java
@@ -43,7 +43,6 @@ import androidx.core.content.res.ResourcesCompat;
 
 import org.drinkless.td.libcore.telegram.TdApi;
 import org.thunderdog.challegram.BaseActivity;
-import org.thunderdog.challegram.Log;
 import org.thunderdog.challegram.R;
 import org.thunderdog.challegram.U;
 import org.thunderdog.challegram.config.Config;
@@ -1245,10 +1244,10 @@ public class MediaWrapper implements FileProgressComponent.SimpleListener, FileP
           twAlternativeHeader = textShort;
           text = Lang.getString(R.string.ProcessingMedia, textShort);
         } else {
-          int progressSize = fileProgress.getProgressSize();
-          int totalSize = fileProgress.getTotalSize();
+          long progressSize = fileProgress.getProgressSize();
+          long totalSize = fileProgress.getTotalSize();
           if (progressSize <= totalSize) {
-            float ratio = (float) progressSize / (float) totalSize;
+            float ratio = (float) ((double) progressSize / (double) totalSize);
             text = Lang.getString(R.string.format_uploadStatus, (int) Math.floor(100f * ratio), textShort);
             twAlternativeHeader = Lang.getString(R.string.format_percentage, (int) Math.floor(100f * ratio));
           } else {
